+++ lib/src/utils/json_config.dart
@@ -1,0 +1,22 @@
+import 'dart:convert';
+import 'dart:io';
+
+import 'package:flutter/services.dart';
+
+/// Get config file as json
+Future<Map<String, String>> getJsonConfig() async {
+  final configString = await rootBundle.loadString('assets/config.json');
+  final jsonConfig = jsonDecode(configString) as Map<String, dynamic>;
+  return jsonConfig.cast<String, String>();
+}
+
+/// Reset config file
+Future<void> resetJsonConfig() async {
+  final jsonConfig = await getJsonConfig();
+  final configPath = '${jsonConfig['app_path']!}/assets/config.json';
+
+  jsonConfig['launch_mode'] = '_${jsonConfig['launch_mode']}';
+  const encoder = JsonEncoder.withIndent('  ');
+  final jsonString = encoder.convert(jsonConfig);
+  File(configPath).writeAsStringSync(jsonString);
+}